General project clean-up
- Removed orphaned old dev-documentation
- Rearranged src files belonging to stlink-lib
- Revised & sorted option byte source code
- Updated devices_boards.md
