--- doc/dev/app-example/README.md
@@ -1,2 +1,0 @@
-This is a simple standalone application example that uses libstlink.
-It can be used as a boilerplate for app development.
