+++ src/st-util/gdb-server.c
@@ -28,6 +28,7 @@
 #include <stlink.h>
 #include <helper.h>
 #include <logging.h>
+#include "flashloader.h"
 #include "gdb-remote.h"
 #include "gdb-server.h"
 #include "semihosting.h"
