+++ src/stlink-lib/calculate.c
