+++ src/stlink-lib/calculate.h
