+++ src/stlink-lib/common.c
