+++ src/stlink-lib/common.h
