+++ src/stlink-lib/flashloader.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
 #include <string.h>
 #include <stlink.h>
+#include "flashloader.h"
 #include "common_flash.h"
 
 #define L1_WRITE_BLOCK_SIZE 0x80
